Vehicle 2 longitudinal and lateral control structure updated.
VehicleSafetyFilter is added as a Simulink Class and will be used to make vehicle stop when collisions happen or destination reached. Apart from that, it will be useful in the future for the implementation of the Robust Planner, taking it as an argument and overriding the reference acceleration and steering angle commands.